third_party/blink/web_tests/webaudio/AudioParam/value-setter-warnings-expected.txt

CONSOLE ERROR: [audit.js] this test requires the explicit comparison with the expected result when it runs with run_web_tests.py.
CONSOLE MESSAGE: Test: test-0
CONSOLE MESSAGE: Test: test-1
CONSOLE MESSAGE: Test: test-2
CONSOLE MESSAGE: Test: test-3
CONSOLE MESSAGE: Test: test-4
CONSOLE MESSAGE: Test: test-5
This is a testharness.js-based test.
Harness: the test ran to completion.
